Enable enhanced CloudWatch metrics in Taupage
This requires setting up an IAM Instance Profile (which requires a
Role, which in turn requires a Policy).

The IAM stuff is global, and replicated across the AWS regions with
eventual consistency.  Beware of run_instances() complaining about
invalid Instance Profile ARN.  We "fix" that with a simple sleep for
30 seconds.